output saturation for narrowband (need to do the same for wideband)
git-svn-id: http://svn.xiph.org/trunk/speex@5445 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ libspeex/testenc.c
@@ -105,8 +105,8 @@ int main(int argc, char **argv)
    {
 	float s=0, e=0;
         for (i=0;i<FRAME_SIZE;++i) {
-            s += in_short[i] * in_short[i];
-            e += (in_short[i]-out_short[i]) * (in_short[i]-out_short[i]);
+            s += (float)in_short[i] * in_short[i];
+            e += ((float)in_short[i]-out_short[i]) * ((float)in_short[i]-out_short[i]);
         }
 	seg_snr += 10*log10((s+1)/(e+1));
 	sigpow += s;
